extract SubmittedComponent
